Separate python projects component
